+++ openhands/server/app.py
@@ -20,7 +20,7 @@ from openhands.server.routes.conversation import app as conversation_api_router
 from openhands.server.routes.feedback import app as feedback_api_router
 from openhands.server.routes.files import app as files_api_router
 from openhands.server.routes.github import app as github_api_router
-from openhands.server.routes.new_conversation import app as new_conversation_api_router
+from openhands.server.routes.manage_conversations import app as new_conversation_api_router
 from openhands.server.routes.public import app as public_api_router
 from openhands.server.routes.security import app as security_api_router
 from openhands.server.routes.settings import app as settings_router
